requirements.txt
dearpygui==2.2
evdev==1.9.3
pynput==1.8.1
python-xlib==0.33
pyusb==1.3.1
six==1.17.0
